requirements.txt
Flask
flask-cors
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth
python-dotenv

